Fix: 머지 후 발생한 레이아웃 bug fix
